SBT build related stuff.
- Remove jerkson, codehaus jackson, and "Json", which have all been deprecated
  for a few releases now.

- Switch scala 2.9 version to 2.9.1, because that's what the maven artifact was,
  and it will make updating less painful.

- Remove non-artifactory resolvers, because we don't want to pull non-cached
  stuff during our builds.

- Execute tests serially. Not sure if this will matter, but the old runner
  did things this way.

- Set test timezone to UTC. This *does* matter because the Intervals test
  assumes it is running in UTC.

- Switch all tests to run like standard JUnit tests (un-nest the classes).
  We are still using simplespec for the matchers syntax, though.
